+++ lib/pages/login_page.dart
@@ -7,7 +7,7 @@ import 'package:pr0gramm/api/preferences.dart';
 import 'package:pr0gramm/api/response_parser.dart';
 import 'package:pr0gramm/content/pr0gramm_content.dart';
 import 'package:pr0gramm/content/pr0gramm_login.dart';
-import 'package:pr0gramm/content/thumb_widget.dart';
+import 'package:pr0gramm/widgets/thumb_widget.dart';
 import 'package:pr0gramm/design/pr0gramm_colors.dart';
 import 'package:pr0gramm/pages/main_page.dart';
 import 'package:pr0gramm/widgets/Design/Pr0Text.dart';
