+++ lib/widgets/thumb_widget.dart
